+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## dev
+  * Enhancements
+  * [Generators] Adding suggestion to cake-php framework;
+
 ## v0.16.1 - (2015-11-08)
 
 * Enhancements